+++ packages/graphql/src/index.js
@@ -1,4 +1,4 @@
-import { createFilter } from 'rollup-pluginutils';
+import { createFilter } from '@rollup/pluginutils';
 import loader from 'graphql-tag/loader';
 
 import toESModules from './toESModules';
@@ -16,9 +16,14 @@ export default function graphql({ include, exclude } = {}) {
       if (!filterExt.test(id)) return null;
 
       // XXX: this.cachable() in graphql-tag/loader
-      const code = toESModules(loader.call({
-        cacheable() { }
-      }, source));
+      const code = toESModules(
+        loader.call(
+          {
+            cacheable() {}
+          },
+          source
+        )
+      );
 
       const map = { mappings: '' };