+++ packages/graphql/tests/samples/basic/index.js
@@ -1,1 +1,0 @@
-export { default as doc } from './basic.graphql';