+++ packages/graphql/CHANGELOG.md
@@ -1,9 +1,23 @@
-# rollup-plugin-json changelog
+# @rollup/plugin-graphql ChangeLog
 
-## 0.1.0
+## v1.0.0
+
+_2019-02-18_
+
+### Features
+
+- Add support for multiple exports
+
+## v0.1.0
+
+_2017-01-16_
+
+### Features
 
 - Use `graphql-tag/loader` directly [PR #3](https://github.com/kamilkisiela/rollup-plugin-graphql/pull/3)
 
-## 0.0.1
+## v0.0.1
+
+_2017-01-07_
 
 - First release